DRT: move CostCalculationStrategy outside InsertionCostCalculator
