+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/insertion/CostCalculationStrategy.java
@@ -1,0 +1,68 @@
+/*
+ * *********************************************************************** *
+ * project: org.matsim.*
+ * *********************************************************************** *
+ *                                                                         *
+ * copyright       : (C) 2021 by the members listed in the COPYING,        *
+ *                   LICENSE and WARRANTY file.                            *
+ * email           : info at matsim dot org                                *
+ *                                                                         *
+ * *********************************************************************** *
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *   See also COPYING, LICENSE and WARRANTY file                           *
+ *                                                                         *
+ * *********************************************************************** *
+ */
+
+package org.matsim.contrib.drt.optimizer.insertion;
+
+import org.matsim.contrib.drt.passenger.DrtRequest;
+
+/**
+ * @author Michal Maciejewski (michalm)
+ */
+public interface CostCalculationStrategy {
+	double calcCost(DrtRequest request, InsertionGenerator.Insertion insertion, double vehicleSlackTime,
+			InsertionCostCalculator.DetourTimeInfo detourTimeInfo);
+
+	class RejectSoftConstraintViolations implements CostCalculationStrategy {
+		@Override
+		public double calcCost(DrtRequest request, InsertionGenerator.Insertion insertion, double vehicleSlackTime,
+				InsertionCostCalculator.DetourTimeInfo detourTimeInfo) {
+			double totalTimeLoss = detourTimeInfo.getTotalTimeLoss();
+			if (totalTimeLoss - vehicleSlackTime > 0
+					|| detourTimeInfo.departureTime - request.getLatestStartTime() > 0
+					|| detourTimeInfo.arrivalTime - request.getLatestArrivalTime() > 0) {
+				return InsertionCostCalculator.INFEASIBLE_SOLUTION_COST;
+			}
+
+			return totalTimeLoss;
+		}
+	}
+
+	class DiscourageSoftConstraintViolations implements CostCalculationStrategy {
+		//XXX try to keep penalties reasonably high to prevent people waiting or travelling for hours
+		//XXX however, at the same time prefer max-wait-time to max-travel-time violations
+		static final double MAX_WAIT_TIME_VIOLATION_PENALTY = 1;// 1 second of penalty per 1 second of late departure
+		static final double MAX_TRAVEL_TIME_VIOLATION_PENALTY = 10;// 10 seconds of penalty per 1 second of late arrival
+
+		@Override
+		public double calcCost(DrtRequest request, InsertionGenerator.Insertion insertion, double vehicleSlackTime,
+				InsertionCostCalculator.DetourTimeInfo detourTimeInfo) {
+			double totalTimeLoss = detourTimeInfo.getTotalTimeLoss();
+			if (totalTimeLoss - vehicleSlackTime > 0) {
+				return InsertionCostCalculator.INFEASIBLE_SOLUTION_COST;
+			}
+
+			double waitTimeViolation = Math.max(0, detourTimeInfo.departureTime - request.getLatestStartTime());
+			double travelTimeViolation = Math.max(0, detourTimeInfo.arrivalTime - request.getLatestArrivalTime());
+			return MAX_WAIT_TIME_VIOLATION_PENALTY * waitTimeViolation
+					+ MAX_TRAVEL_TIME_VIOLATION_PENALTY * travelTimeViolation
+					+ totalTimeLoss;
+		}
+	}
+}
